\name{BCfireArea}
\alias{BCfireArea}
\title{BC Fire Area}
\usage{data(BCfireArea)}
\description{
The \code{BCfireArea} time series object consists of 13 observations on annual area burnt in the province of BC.
}
\format{
  A time series object
}
\examples{
  ts.plot(BCfireArea)
}
\keyword{datasets}
