\name{longitudinalAcceleration}
\alias{longitudinalAcceleration}
\title{Longitudinal Acceleration Measurements on an Air Tanker}
\usage{data(longitudinalAcceleration)}
\description{
Longitudinal acceleration measurements of an air tanker fighting a forest
wildfire taken at 1 second intervals. 
}
\format{
  A time series object
}
\examples{
    acf(longitudinalAcceleration)
}
\keyword{datasets}
