\name{nickel}
\alias{nickel}
\title{Electroless nickel concentrations}
\usage{data(nickel)}
\description{
Electroless nickel concentrations in a chrome plating process were measured at the beginning of each eight hour work shift for a period of 25 days.  A  concentration of 4.5 ounces per gallon is considered optimal in this application.
}
\format{
  A time series object
}
\source{Farnum, N. (1994) Statistical Quality Control and Improvement. 
Belmont, Duxbury Press.
}
\examples{
    ts.plot(nickel)
}
\keyword{datasets}
